Task 2 was much easier.. Hope it last
I had the wrong idea at the start - was thinking about doing the same setup as task 1 but quickly realized it wont work to identify X patterns. Once I figure that out the solution came relatively quickly. For some reason it didnt work when I was using the if check for 'A' inside the xSearch function. Could have been that or the other checks I removed for bound checking.