Solution for day 5 task 2
This is as close as I can get to getting a working solution to this problem. Not sure where my logic is wrong. I changed the logic to only add the middle element of the pages when it was reordered, and only after no more swapps occur per line. Will be seeking help